fix: the four things opening the security domain found
**Four orientations nobody could be reviewed for.** The domain already had
`security-engineer`, `pentester-web`, `pentester-mobile` and `soc-analyst`,
each with a null `reviewer_group` — visible rather than silently open, as the
column's comment says, which in practice meant nobody could review the work.
All four are archived and pointed at what replaces them, and a guard in the
migration refuses to leave a live curated security orientation without a review
family. `pentester-web` and `pentester-mobile` named a target rather than a
craft; those are tags on the red team row now.

**An NDA nobody could name.** `missions.nda_required` had existed for months
with no column saying which agreement, so the flag meant "you must sign
something". Creation now names the standard mutual agreement by default, and a
client's own document has to be uploaded rather than linked — an agreement this
platform cannot read is one it cannot hash.

**The first round of questions on any finding returned a 500.** The next round
number came back as `max(round_no) + 1`, which PostgreSQL widens to `int4`,
decoded into an `i16`. Cast at the query.

**security.txt pointed at a page that does not exist.** `Policy:` and
`Acknowledgments:` were written before the programme had a scope page or a hall
of fame. A scanner reports a dead policy line as present, which is worse than
absent.

Also: the orientation cap-of-three test asserted its 400 using `pentester-web`,
which the migration above archives — it would have kept passing for the wrong
reason.

+++ src/services/security_findings.rs
@@ -962,9 +962,12 @@ pub async fn open_round(
         )));
     }
 
+    // `::SMALLINT` on the way out, because `max(smallint) + 1` is an integer in
+    // PostgreSQL and sqlx does not narrow: without the cast the decode fails and
+    // the endpoint answers 500 to every first round.
     let next: i16 = sqlx::query_scalar(
-        "SELECT COALESCE(max(round_no), 0) + 1 FROM security_finding_rounds
-          WHERE finding_id = $1",
+        "SELECT (COALESCE(max(round_no), 0) + 1)::SMALLINT
+           FROM security_finding_rounds WHERE finding_id = $1",
     )
     .bind(finding_id)
     .fetch_one(db)
